// apps/web/src/api/generated/model/accessTokenResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Starter API
 * OpenAPI spec version: 0.1.0
 */
import type { AccessTokenResponseTokenType } from './accessTokenResponseTokenType';

export interface AccessTokenResponse {
  accessToken: string;
  tokenType: AccessTokenResponseTokenType;
  expiresAt: string;
}